Remove optimism explorer redux (#4524)
* Under construction: removal of fallbackExplorer

* Remove forceFallbackProvider debug flag

* Fix typo in logs

* Replace previous fetchOnchainBalances usage with fetchAssetsFromRefraction

fetchAssetsFromRefraction is a `get` action to the
address assets endpoint and will be a good substitute
now that fetchOnchainBalances has been removed.

* Replace fetchOnchainBalances with useRefreshAccountData in useAccountRefresh hook

* Remove unused migratedTokens reference

* Remove unused covalent handler

* Remove unused Covalent ETH address

* Add ETH mainnet to chain assets reference for hardhat

* Fetch balances for hardhat or testnet

* Remove optimismExplorerInit in explorer redux

Since we now have backend support for optimism
the same way we have for other L2s, we can remove the
optimismExplorer. With this removed, the explorerInitL2
in the explorer redux does not need to be defined and upstream
can call on fetchAssetsFromRefraction directly

* Remove optimismExplorer clear state from useResetAccountState

* Remove optimismExplorer

* Remove unused balance checker contract abi ovm from references
