+++ src/hooks/useRefreshAccountData.ts
@@ -3,7 +3,7 @@ import delay from 'delay';
 import { useCallback, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import NetworkTypes from '../helpers/networkTypes';
-import { fetchOnchainBalances } from '../redux/fallbackExplorer';
+import { fetchAssetsFromRefraction } from '../redux/explorer';
 import { uniqueTokensRefreshState } from '../redux/uniqueTokens';
 import { updatePositions } from '../redux/usersPositions';
 import { walletConnectLoadState } from '../redux/walletconnect';
@@ -33,9 +33,7 @@ export default function useRefreshAccountData() {
         ? dispatch(fetchWalletENSAvatars())
         : null;
       const getUniqueTokens = dispatch(uniqueTokensRefreshState());
-      const balances = dispatch(
-        fetchOnchainBalances({ keepPolling: false, withPrices: false })
-      );
+      const balances = dispatch(fetchAssetsFromRefraction());
       const wc = dispatch(walletConnectLoadState());
       const uniswapPositions = dispatch(updatePositions());
       return Promise.all([
