+++ src/references/chain-assets.json
@@ -14,7 +14,25 @@
         },
         "symbol": "ETH"
       },
-      "quantity": 0
+      "quantity": "0"
+    }
+  ],
+  "mainnet": [
+    {
+      "asset": {
+        "asset_code": "eth",
+        "coingecko_id": "ethereum",
+        "decimals": 18,
+        "icon_url": "https://s3.amazonaws.com/icons.assets/ETH.png",
+        "name": "Ether",
+        "price": {
+          "changed_at": 1582568575,
+          "relative_change_24h": -4.586615622469276,
+          "value": 259.2
+        },
+        "symbol": "ETH"
+      },
+      "quantity": "0"
     }
   ],
   "optimism": []
