--- src/references/migratedTokens.ts
@@ -1,6 +1,0 @@
-export default {
-  '0xc011a72400e58ecd99ee497cf89e3775d4bd732f':
-    '0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f', // SNX
-  '0x57ab1e02fee23774580c119740129eac7081e9d3':
-    '0x57ab1ec28d129707052df4df418d58a2d46d5f51', // SUSD
-};
